+++ test/formatters/samplecode/ktlint/in.kt
@@ -1,0 +1,8 @@
+class Test
+
+{
+fun test(  v: String )
+{
+ println(v)
+}
+}